+++ src/Fable.Transforms/Replacements.Util.fs
@@ -557,7 +557,10 @@ let partialApplyAtRuntime (com: Compiler) t arity (expr: Expr) (partialArgs: Exp
             let curriedType = makeLambdaType argTypes returnType
 
             let curried =
-                Helper.LibCall(com, "Util", $"curry{argTypes.Length}", curriedType, [ expr ])
+                match com.Options.Language with
+                | Python ->
+                    Helper.LibCall(com, "Curry", "curry", curriedType, [ makeNativeIntConst argTypes.Length; expr ])
+                | _ -> Helper.LibCall(com, "Util", $"curry%d{argTypes.Length}", curriedType, [ expr ])
 
             match partialArgs with
             | [] -> curried
@@ -608,7 +611,9 @@ let uncurryExprAtRuntime (com: Compiler) arity (expr: Expr) =
         | Python ->
             let uncurriedType = DelegateType(argTypes, returnType)
 
-            Helper.LibCall(com, "Util", $"uncurry{arity}", uncurriedType, [ expr ])
+            match com.Options.Language with
+            | Python -> Helper.LibCall(com, "Curry", "uncurry", uncurriedType, [ makeNativeIntConst arity; expr ])
+            | _ -> Helper.LibCall(com, "Util", $"uncurry%d{arity}", uncurriedType, [ expr ])
         | _ ->
             // let makeArgIdent typ = makeTypedIdent typ $"a{com.IncrementCounter()}$"
             // let argIdents = argTypes |> List.map makeArgIdent