+++ src/fable-library-py/fable_library/task_builder.py
@@ -8,11 +8,11 @@ from typing import (
 )
 
 from .task import from_result, zero
-from .util import UNIT, IDisposable
+from .util import UNIT, IDisposable, Unit
 
 
 class Delayed[T](Protocol):
-    def __call__(self, __unit=UNIT) -> Awaitable[T]: ...
+    def __call__(self, __unit: Unit = UNIT) -> Awaitable[T]: ...
 
 
 class TaskBuilder: